+++ lwprintf/src/include/lwprintf/lwprintf_opt.h
@@ -58,7 +58,7 @@ extern "C" {
  * \note            When `LWPRINTF_CFG_OS` is enabled, user must implement functions in \ref LWPRINTF_SYS group.
  */
 #ifndef LWPRINTF_CFG_OS
-#define LWPRINTF_CFG_OS                            0
+#define LWPRINTF_CFG_OS                             0
 #endif
 
 /**
@@ -69,7 +69,24 @@ extern "C" {
  *                  definition before you define handle type
  */
 #ifndef LWPRINTF_CFG_OS_MUTEX_HANDLE
-#define LWPRINTF_CFG_OS_MUTEX_HANDLE               void *
+#define LWPRINTF_CFG_OS_MUTEX_HANDLE                void *
+#endif
+
+/**
+ * \brief           Enables `1` or disables `0` manual mutex lock.
+ * 
+ * When this feature is enabled, together with \ref LWPRINTF_CFG_OS, behavior is as following:
+ * - System mutex is kept created
+ * - Calls to direct printing functions are not thread-safe by default anymore
+ * - Calls to sprintf (buffer functions) are kept thread-fe
+ * - User must manually call \ref lwprintf_protect or \ref lwprintf_protect_ex functions to protect direct printing operation
+ * - User must manually call \ref lwprintf_unprotect or \ref lwprintf_unprotect_ex functions to exit protected area
+ * 
+ * \note            If you prefer to completely disable locking mechanism with this library,
+ *                  turn off \ref LWPRINTF_CFG_OS and fully manually handle mutual exclusion for non-reentrant functions
+ */
+#ifndef LWPRINTF_CFG_OS_MANUAL_PROTECT              
+#define LWPRINTF_CFG_OS_MANUAL_PROTECT              0
 #endif
 
 /**